+++ packages/manifest/src/index.ts
@@ -21,6 +21,8 @@ export const initState: DataInitializer = (): ElementData => ({
 // Can be loaded from package.json
 export const version = '1.0';
 
+export const isEmpty = (data: ElementData): boolean => !data.playbackId;
+
 // UI configuration for Tailor CMS
 const ui = {
   // Display icon, https://pictogrammers.com/library/mdi/
@@ -36,6 +38,7 @@ const manifest: ElementManifest = {
   name,
   ssr: false,
   initState,
+  isEmpty,
   ui,
 };